requirements.txt
# AWS SDK
boto3==1.26.137
botocore==1.29.137

# OpenAI
openai==1.3.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Image Processing
Pillow==10.0.0
opencv-python-headless==4.8.0.74

# Audio Processing
sounddevice==0.4.6
numpy==1.24.3

# Storage
redis==4.5.4

# Utils
python-dotenv==1.0.0
pydantic==2.4.2
PyYAML==6.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0

# Monitoring
prometheus-client==0.17.1

# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0